Reorganize AuthModal components into designated folder
