+++ src/pages/Landing/AuthModal/AuthModal.module.css
@@ -1,9 +1,3 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
 .modalOverlay {
   position: fixed;
   inset: 0;
@@ -93,9 +87,19 @@
   color: black;
 }
 
+.form input:focus {
+  margin: 0;
+  outline: none;
+  box-shadow: none;
+}
+
 .errorText {
   color: red;
   font-size: 1rem;
+  line-height: 0.5rem;
+  margin: 0;
+  padding: 0;
+  font-size: 0.8rem;
 }
 
 .submitButton {
